delete geofences jobintent, leave only broadcast receiver
