+++ app/src/main/AndroidManifest.xml
@@ -141,11 +141,6 @@
             android:enabled="true"
             android:exported="true" />
 
-        <service
-            android:name=".services.GeofenceTransitionsJobIntentService"
-            android:exported="true"
-            android:permission="android.permission.BIND_JOB_SERVICE" />
-
     </application>
 
 </manifest>
